Fixed id number issue, requires further consideration
The IDs are being generated from the android id which is guaranteed to be
unique. However, the android ID is not guaranteed to fit inside an int.
This was leading to a NumberFormatException. Now this has been fixed
with BigInteger but the id is no longer guaranteed to be unique (used
the modulus of the android id). Needs further consideration in future.